# ggml: Support OpenMP for multi-thread processing