Currently, ggml multi-thread processing is  implemented with `pthread_create` and `pthread_join`. This is simple and straightforward, but has the following issues

- Difficult to analyze performance per TID (e.g., `perf`) because each thread is created each time a token is generating.
- Thread affinity and core binding settings are not flexible.

Therefore, I suggest an option using OpenMP, which is general multi-threaded library. it is optional, and by default the previous implementation works.

Below are the results measured on an AWS m7i.24xlarge (Intel Xeon 8488).

```bash
### build
$ cmake -DLLAMA_OPENMP=ON -B build -S .
$ cmake --build build -j$(($(nproc)/2))

### run
$ export OMP_PROC_BIND=TRUE # if need
$ main --model models/llama-2-7b-chat.Q4_0.gguf --temp 0.1 --threads XX --prompt 'AI is going to' --n-predict 256 --seed 0
```

|   Threads   | Pthread<br />[token/sec] | OpenMP<br />[token/sec] | OpenMP<br />OMP_PROC_BIND=TRUE<br />[token/sec] |
| ---- | ------------------------ | ------------------------------------------------ | ----------------------------------------------- |
| 6    | 11.07                    | 11.24                                            | 7.11                                            |
| 12   | 19.28                    | 19.62                                            | 13.37                                           |
| 24   | 26.81                    | 27.11                                            | 22.92                                           |
| 36   | 26.73                    | 28.94                                            | 28.76                                           |
| 42   | 26.08                    | 27.21                                            | 30.54                                           |
| 48   | 25.18                    | 26.72                                            | 30.97                                           |

![image](https://github.com/ggerganov/llama.cpp/assets/62578291/cb2becb3-e9d9-41e7-b0cf-9ac4e2cc5301)

OpenMP version has slightly better performance than the pthread implementation. And setting OMP_PROC_BIND=TRUE improves performance on many threads (Threads >~ 36).

It would be good for users to have additional OpenMP options to tuning for their system.
Any comments are welcome!

related : #7342 #7526