@slaren 2024-05-29
Here is the same test with NKVO that I have been running in the other thread pools PRs:

| GPU         | Model         |   NKVO | Test   |   t/s master |   t/s feat-openmp-support-draft |   Speedup |
|:------------|:--------------|-------:|:-------|-------------:|--------------------------------:|----------:|
| RTX 3090 Ti | llama 7B Q4_0 |      0 | pp512  |      4539.81 |                         4534.23 |      1.00 |
| RTX 3090 Ti | llama 7B Q4_0 |      0 | tg128  |       153.90 |                          155.25 |      1.01 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | pp512  |       502.87 |                          527.40 |      1.05 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | tg128  |        20.85 |                           62.17 |      2.98 |

I have also tested with CPU only, on native Windows with MSVC, and on a Apple MBP with an M3 Max, and I haven't found any performance regressions. So I have to say congratulations, this is the first thread pool attempt that achieves a very large reduction in the thread launch overhead, without any significant downsides. I think we could use OMP by default.

Some concerns:
- The value of `num_threads` is not always respected by OMP, and when this happens the result is a deadlock. The [OMP logic](https://www.openmp.org/spec-html/5.0/openmpsu35.html#x55-880002.6.1) is fairly complicated and it may depend on factors outside of our control. We may need to update the number of threads in the shared state according to the value returned by `omp_get_num_threads`.
- It may be better to avoid using OMP at all if `n_threads == 1`, and instead use the main thread. The OMP implementations that I tested seem to be smart enough to do this already, but I don't think this is guaranteed, and I it would be more reliable to handle this case ourselves and avoid the OMP overhead entirely.