@netrunnereve 2024-05-30
> I have also tested with CPU only, on native Windows with MSVC, and on a Apple MBP with an M3 Max, and I haven't found any performance regressions. So I have to say congratulations, this is the first thread pool attempt that achieves a very large reduction in the thread launch overhead, without any significant downsides. I think we could use OMP by default.

On an old 4 core/8 thread Xeon v2 there's a small regression with four threads but it's the same as master with eight. Typically with these threadpool implementations the speedup only happens when you use a lot of threads and thus have to deal with the overhead.
`master`
| model                          |       size |     params | backend    |    threads |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ---------------: |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          4 |         pp512 |      4.38 ± 0.00 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          4 |         tg128 |      3.51 ± 0.00 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          8 |         pp512 |      4.64 ± 0.02 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          8 |         tg128 |      3.52 ± 0.00 |

`PR`
| model                          |       size |     params | backend    |    threads |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ---------------: |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          4 |         pp512 |      4.34 ± 0.00 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          4 |         tg128 |      3.30 ± 0.08 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          8 |         pp512 |      4.64 ± 0.00 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          8 |         tg128 |      3.52 ± 0.00 |

`PR with OMP_PROC_BIND=TRUE`
| model                          |       size |     params | backend    |    threads |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ---------------: |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          8 |         pp512 |      4.64 ± 0.01 |
| llama 7B Q6_K                  |   5.53 GiB |     7.24 B | CPU        |          8 |         tg128 |      3.49 ± 0.00 |