@slaren 2024-05-30
I tried with a very small stories260k model, since that should exacerbate the overhead of starting the threads, but it is still consistently faster with this PR (except 1 threads). Is it really a given that waking up a thread needs to be more expensive than starting a thread?

| CPU           | Model                      |   Threads | Test   |   t/s master |   t/s omp |   Speedup |
|:--------------|:---------------------------|----------:|:-------|-------------:|--------------------------------:|----------:|
|  i9-13900K | stories260k F32 |         1 | pp512  |      6920.29 |                         6882.03 |      0.99 |
|  i9-13900K | stories260k F32 |         1 | tg128  |      6937.02 |                         6865.42 |      0.99 |
|  i9-13900K | stories260k F32 |         2 | pp512  |     12365.50 |                        12665.08 |      1.02 |
|  i9-13900K | stories260k F32 |         2 | tg128  |      4687.61 |                         5891.87 |      1.26 |
|  i9-13900K | stories260k F32 |         3 | pp512  |     17501.55 |                        17597.04 |      1.01 |
|  i9-13900K | stories260k F32 |         3 | tg128  |      3988.88 |                         5272.38 |      1.32 |
|  i9-13900K | stories260k F32 |         4 | pp512  |     22587.09 |                        22778.87 |      1.01 |
|  i9-13900K | stories260k F32 |         4 | tg128  |      3454.17 |                         4633.26 |      1.34 |
|  i9-13900K | stories260k F32 |         5 | pp512  |     27142.85 |                        27477.98 |      1.01 |
|  i9-13900K | stories260k F32 |         5 | tg128  |      3028.37 |                         4052.77 |      1.34 |
|  i9-13900K | stories260k F32 |         6 | pp512  |     30735.90 |                        31353.86 |      1.02 |
|  i9-13900K | stories260k F32 |         6 | tg128  |      2669.83 |                         3668.35 |      1.37 |
|  i9-13900K | stories260k F32 |         7 | pp512  |     31723.15 |                        35381.52 |      1.12 |
|  i9-13900K | stories260k F32 |         7 | tg128  |      2388.36 |                         3295.09 |      1.38 |
|  i9-13900K | stories260k F32 |         8 | pp512  |     32668.71 |                        31819.11 |      0.97 |
|  i9-13900K | stories260k F32 |         8 | tg128  |      2140.98 |                         2944.40 |      1.38 |